Extract animations and lighting demos
