--- DX12Renderer/Shaders/AnimationsDemo_Bone_PS.hlsl
@@ -1,4 +1,0 @@
-float4 main() : SV_Target
-{
-    return float4(1.0, 0.0, 0.0, 0.5);
-}
