--- DX12Renderer/include/TaaCBuffer.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <DirectXMath.h>
-
-struct TaaCBuffer
-{
-	DirectX::XMMATRIX PreviousModelViewProjectionMatrix;
-
-	DirectX::XMFLOAT2 JitterOffset;
-	float _Padding[2];
-};
-
