+++ Demos/LightingDemo/CMakeLists.txt
@@ -1,8 +1,8 @@
 ﻿cmake_minimum_required(VERSION 3.8.0)
 
-project("DX12Renderer")
+project("LightingDemo")
 
-set(TARGET_NAME DX12Renderer)
+set(TARGET_NAME LightingDemo)
 include(${CMAKE_SOURCE_DIR}/CMakeIncludes/ProjectBase.cmake)
 include(${CMAKE_SOURCE_DIR}/CMakeIncludes/DemoBase.cmake)
