+++ Demos/LightingDemo/include/ParticleSystemPso.h
