+++ Demos/LightingDemo/include/PointLightPso.h
