+++ Demos/LightingDemo/include/Scene.h
