+++ Demos/LightingDemo/include/SceneRenderer.h
