+++ Demos/LightingDemo/shaders/LightingDemo_Blit_VS.hlsl
