+++ Demos/LightingDemo/shaders/LightingDemo_Bloom_Downsample_PS.hlsl
