+++ Demos/LightingDemo/shaders/LightingDemo_PointLight_VS.hlsl
