+++ Demos/LightingDemo/shaders/LightingDemo_ShadowCaster_VS.hlsl
