+++ Demos/LightingDemo/src/main.cpp
@@ -1,4 +1,3 @@
-#include <LightingDemo/LightingDemo.h>
-
+#include <LightingDemo.h>
 #define DEMO_TYPE LightingDemo
 #include <Framework/DemoMain.h>
